Split native shell motion overlay
